FISH/requirements.txt
playwright>=1.40.0
pyyaml>=6.0
python-dotenv>=1.0.0
schedule>=1.2.0
loguru>=0.7.0
